% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizar.R
\name{normalizar}
\alias{normalizar}
\title{Normalize strings}
\usage{
normalizar(x)
}
\arguments{
\item{x}{A string.}
}
\value{
The normalized string
}
\description{
Normalize strings, removing diacritics, numbers, punctuations, leading and
 trailing blank spaces and double spaces.
}
\examples{
normalizar("São Paulo 2020!")
}
